+++ lib/packages/wikibase/data-model/src/Statement/StatementList.php
@@ -342,6 +342,15 @@ class StatementList implements IteratorAggregate, Comparable, Countable {
 	}
 
 	/**
+	 * Removes all statements from this list.
+	 *
+	 * @since 6.0
+	 */
+	public function clear() {
+		$this->statements = [];
+	}
+
+	/**
 	 * @see http://php.net/manual/en/language.oop5.cloning.php
 	 *
 	 * @since 5.1